commit hashtags comparsion

+++ src/Shell.hs
@@ -10,9 +10,13 @@ module Shell
     gentooSync
   ) where
 
+import Data.List.Split
+
 import System.Info (os)
 import System.Directory
 import System.FilePath((</>))
+
+import System.Process
 
 import Control.Monad
 import Control.Eternal
@@ -22,12 +26,21 @@ rebasefork path branch upstream =
     doesDirectoryExist path                  >>= \dirExist ->
         doesDirectoryExist <| path </> ".git" >>= \git ->
          if git then if dirExist
-                        then do exc path $ "git checkout "    ++ branch
+                        then setCurrentDirectory path >> do
+                                exec $ "git checkout " ++ branch
                                     ++ " & git reset --hard"
-                                    ++ " & git rebase --abort & git pull origin "   ++ branch
-                                    ++ " & git fetch "                              ++ upstream
-                                    ++ " & git pull --rebase "                      ++ upstream
-                                    ++ " & git push --force origin "                ++ branch
+                                    ++ " & git rebase --abort"
+                                loc <- readProcess "git" ["log", "-n", "1", "--pretty=format:%H"] []
+                                rem <- readProcess "git" (["ls-remote"] ++ (splitOn " " upstream)) []
+                                let remote = (splitOn "\t" rem) !! 0
+                                putStrLn $ "Local: "  ++ loc
+                                putStrLn $ "Remote: " ++ remote
+                                if  remote == loc
+                                    then putStrLn $ path ++ " is up to date"
+                                    else exec $ "git pull origin "             ++ branch
+                                              ++ " & git fetch "               ++ upstream
+                                              ++ " & git pull --rebase "       ++ upstream
+                                              ++ " & git push --force origin " ++ branch
                                 return True     -- rebase complete
                         else    return False    -- directory doesn't exist
                 else            return True     -- directory exists but it's not a git
